Struct spirv_cross::spirv::Ast

pub struct Ast<TTarget> where
    TTarget: Target,  { /* fields omitted */ }

An abstract syntax tree that corresponds to a SPIR-V module.

Implementations

impl Ast<Target>

pub fn build_combined_image_samplers(&mut self) -> Result<(), ErrorCode>

pub fn get_combined_image_samplers(
&mut self
) -> Result<Vec<CombinedImageSampler>, ErrorCode>

pub fn add_header_line(&mut self, line: &str) -> Result<(), ErrorCode>

pub fn flatten_buffer_block(&mut self, id: u32) -> Result<(), ErrorCode>

impl Ast<Target>

pub fn set_root_constant_layout(
&mut self,
layout: Vec<RootConstant>
) -> Result<(), ErrorCode>

impl Ast<Target>

pub fn is_rasterization_enabled(&self) -> Result<bool, ErrorCode>

impl<TTarget> Ast<TTarget> where
Self: Parse<TTarget> + Compile<TTarget>,
TTarget: Target,

pub fn get_decoration(
&self,
id: u32,
decoration: Decoration
) -> Result<u32, ErrorCode>

Gets a decoration.

pub fn get_name(&mut self, id: u32) -> Result<String, ErrorCode>

Gets a name. If not defined, an empty string will be returned.

pub fn set_name(&mut self, id: u32, name: &str) -> Result<(), ErrorCode>

Sets a name.

pub fn unset_decoration(
&mut self,
id: u32,
decoration: Decoration
) -> Result<(), ErrorCode>

Unsets a decoration.

pub fn set_decoration(
&mut self,
id: u32,
decoration: Decoration,
argument: u32
) -> Result<(), ErrorCode>

Sets a decoration.

pub fn get_entry_points(&self) -> Result<Vec<EntryPoint>, ErrorCode>

Gets entry points.

pub fn get_cleansed_entry_point_name(
&self,
entry_point_name: &str,
execution_model: ExecutionModel
) -> Result<String, ErrorCode>

Gets cleansed entry point names. compile must be called first.

pub fn get_active_buffer_ranges(
&self,
id: u32
) -> Result<Vec<BufferRange>, ErrorCode>

Gets active buffer ragnes. Useful for push constants.

pub fn get_specialization_constants(
&self
) -> Result<Vec<SpecializationConstant>, ErrorCode>

Gets all specialization constants.

pub fn set_scalar_constant(
&mut self,
id: u32,
value: u64
) -> Result<(), ErrorCode>

Set reference of a scalar constant to a value, overriding the default.

Can be used to override specialization constants.

pub fn get_shader_resources(&self) -> Result<ShaderResources, ErrorCode>

Gets shader resources.

pub fn get_type(&self, id: u32) -> Result<Type, ErrorCode>

Gets the SPIR-V type associated with an ID.

pub fn get_member_name(&self, id: u32, index: u32) -> Result<String, ErrorCode>

Gets the identifier for a member located at index within an OpTypeStruct.

pub fn get_member_decoration(
&self,
id: u32,
index: u32,
decoration: Decoration
) -> Result<u32, ErrorCode>

Gets a decoration for a member located at index within an OpTypeStruct.

pub fn set_member_decoration(
&mut self,
id: u32,
index: u32,
decoration: Decoration,
argument: u32
) -> Result<(), ErrorCode>

Sets a decoration for a member located at index within an OpTypeStruct.

pub fn get_declared_struct_size(&self, id: u32) -> Result<u32, ErrorCode>

Gets the effective size of a buffer block.

pub fn get_declared_struct_member_size(
&self,
id: u32,
index: u32
) -> Result<u32, ErrorCode>

Gets the effective size of a buffer block struct member.

pub fn rename_interface_variable(
&mut self,
resources: &[Resource],
location: u32,
name: &str
) -> Result<(), ErrorCode>

Renames an interface variable.

pub fn get_work_group_size_specialization_constants(
&self
) -> Result<WorkGroupSizeSpecializationConstants, ErrorCode>

Gets work group size specialization constants.

pub fn parse(module: &Module<'_>) -> Result<Self, ErrorCode>

Parses a module into Ast.

pub fn set_compiler_options(
&mut self,
options: &Self::CompilerOptions
) -> Result<(), ErrorCode>

Sets compile options.

pub fn compile(&mut self) -> Result<String, ErrorCode>

Compiles an abstract syntax tree to a String in the specified TTarget language.